+++ prisma/migrations/20251123122634_add_show_purchase_options_to_category/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Category" ADD COLUMN     "showPurchaseOptions" BOOLEAN NOT NULL DEFAULT true;